requirements.txt
Flask==3.1.1
flask-cors==6.0.0
mysql-connector-python==9.3.0
PyGithub==2.6.1
